// packages/wb/src/endpoints/promotion/get-v1-adverts/types.ts
// Generated public types for this Wildberries endpoint.
/**
 * Параметры метода «Список медиакампаний».
 *
 * Метод возвращает список всех [медиакампаний](/openapi/promotion#tag/media/operation/getV1Advert) продавца по их типам и статусам.
 *
 * Лимит запросов (/openapi/api-information#tag/introduction/Limity-zaprosov) на один аккаунт продавца:
 *
 * | Тип | Период | Лимит | Интервал | Всплеск |
 * | --- | --- | --- | --- | --- |
 * | Персональный | 1 сек | 10 запросов | 100 мс | 10 запросов |
 * | Сервисный | 1 сек | 10 запросов | 100 мс | 10 запросов |
 * | Базовый с секретом | 1 сек | 10 запросов | 100 мс | 10 запросов |
 * | Базовый | 1 ч | 1 запрос | 1 ч | 1 запрос |
 *
 * GET /adv/v1/adverts
 *
 * Источник: https://dev.wildberries.ru/docs/openapi/promotion
 */
export type GetV1AdvertsInput = {
  /**
   * Параметры строки запроса.
   */
  readonly query?: {
    /**
     * Статус медиакампании:
     *  - `1` — черновик
     *  - `2` — модерация
     *  - `3` — отклонена (с возможностью вернуть на модерацию)
     *  - `4` — готова к запуску
     *  - `5` — запланирована
     *  - `6` — на показах
     *  - `7` — завершена
     *  - `8` — отменена
     *  - `9` — приостановлена продавцом
     *  - `10` — пауза по дневному лимиту
     *  - `11` — пауза
     *
     * Пример: `"1,3,7"`.
     */
    readonly status?: string;

    /**
     * Тип медиакампании:
     * - `1` — размещение по дням
     * - `2` — размещение по просмотрам
     *
     * Пример: `1`.
     */
    readonly type?: number;

    /**
     * Количество кампаний в ответе
     *
     * Пример: `1`.
     */
    readonly limit?: number;

    /**
     * Смещение относительно первой медиакампании
     *
     * Пример: `1`.
     */
    readonly offset?: number;

    /**
     * Порядок вывода ответа:
     * - `create` — по времени создания медиакампании
     * - `id` — по ID медиакампании
     *
     * Пример: `"id"`.
     */
    readonly order?: string;

    /**
     * Порядок сортировки:
     * - `desc` — от большего к меньшему
     * - `asc` — от меньшего к большему
     *
     * Пример: `"desc"`.
     */
    readonly direction?: string;
  };
};
/**
 * Успешный ответ метода «Список медиакампаний».
 *
 * 200: Успешно.
 *
 * 204: Медиакампании не найдены.
 */
export type GetV1AdvertsResponse =
  | readonly {
      /**
       * ID медиакампании
       */
      readonly advertId?: number;

      /**
       * Название медиакампании
       */
      readonly name?: string;

      /**
       * Название бренда
       */
      readonly brand?: string;

      /**
       * Тип медиакампании:
       * - `1` — размещение по дням
       * - `2` — размещение по просмотрам
       */
      readonly type?: number;

      /**
       * Статус медиакампании:
       *  - `1` — черновик
       *  - `2` — модерация
       *  - `3` — отклонена (с возможностью вернуть на модерацию)
       *  - `4` — готова к запуску
       *  - `5` — запланирована
       *  - `6` — на показах
       *  - `7` — завершена
       *  - `8` — отменена
       *  - `9` — приостановлена продавцом
       *  - `10` — пауза по дневному лимиту
       *  - `11` — пауза
       */
      readonly status?: number;

      /**
       * Время создания медиакампании
       *
       * Формат: `date-time`.
       */
      readonly createTime?: string;

      /**
       * Время завершения медиакампании
       *
       * Формат: `date-time`.
       */
      readonly endTime?: string;
    }[]
  | undefined;